refactor: order, company service 자체 responseDto, ErrorCode 제거 및 controller core ResponseDto 사용 수정
